requirements.txt
appdirs==1.4.3
arrow==0.10.0
beautifulsoup4==4.5.3
billboard.py==3.0.2
lxml==3.7.3
packaging==16.8
pyparsing==2.2.0
python-dateutil==2.6.0
requests==2.13.0
six==1.10.0
